refactor(editor): organize components/editor into role-based subfolders
Group the flat components/editor/ files by role: panel/, canvas/,
preview/, controls/, dialogs/, utils/ (hooks/ already existed).
EditorShell and EditorWorkspaceSection stay at the root as the
public entrypoints. Update import sites and convert one brittle
relative path-data import to an absolute @/lib/... path.
