+++ components/editor/EditorWorkspaceSection.tsx
@@ -4,7 +4,7 @@ import { useCallback, useEffect, type MutableRefObject } from "react";
 import type { EditorAssetSummary } from "@/lib/editor/types";
 import type { CustomizationState } from "@/lib/types";
 import { useEditorDocument } from "@/components/editor/hooks/useEditorDocument";
-import { EditorWorkspacePanel } from "@/components/editor/EditorWorkspacePanel";
+import { EditorWorkspacePanel } from "@/components/editor/panel/EditorWorkspacePanel";
 import { selectAssetInStore } from "@/stores/editor-selection";
 
 interface EditorWorkspaceSectionProps {
