+++ components/editor/preview/EditorSvgPreview.tsx
