+++ components/editor/utils/draw-utils.ts
