+++ components/editor/utils/svg-path-utils.ts
