+++ components/landing/svg/rocket-interactive.tsx
@@ -7,7 +7,7 @@ import {
   pointsToSvgPath,
   addControlPointsToAllAnchors,
   type ControlPoint,
-} from "@/components/editor/svg-path-utils";
+} from "@/components/editor/utils/svg-path-utils";
 
 const ROCKET_PATHS = [
   "M34.6981 57.4011L27.8523 50.5553H16.7764C12.9222 50.5553 10.5417 46.3507 12.5246 43.0459L18.5609 32.9855C21.2491 28.5051 26.091 25.7637 31.3161 25.7637H48.9254C61.401 12.1537 74.9144 2.69453 92.3142 1.18412C95.0423 0.94731 97.2939 3.19859 97.0569 5.92672C95.5466 23.3264 86.0876 36.8401 72.4774 49.3158V66.9248C72.4774 72.1499 69.736 76.9917 65.2556 79.6801L55.1952 85.7164C51.8904 87.6992 47.6858 85.3187 47.6858 81.4646V70.3887L41.1919 63.8949M34.6981 57.4011H24.0005C21.3304 57.4011 19.9907 60.6271 21.8753 62.5186L36.0668 76.7616C37.9551 78.6567 41.1919 77.3194 41.1919 74.6441V63.8949M34.6981 57.4011L41.1919 63.8949",
